+++ R/plotImage.R
@@ -1,8 +1,0 @@
-plotImage <- function(m, x = NULL, y = NULL) {
-  image(
-    x = x,
-    y = y,
-    z = matrix(1:length(m), ncol(m), nrow(m)),
-    col = t(m[nrow(m):1, ])
-  )
-}